final/override:不想被派生类重载则置为virtual并标记final,确定为派生类中的重载版本则置为virtual并标记override

+++ test.sh
@@ -1,3 +1,15 @@
 # !/bin/sh
-rm -f a
-g++ -o a -std=c++11 static_assert.cpp
+if [ $# -ne 1 ]; then
+    echo "Usage: $0 c/b"
+    exit 1;
+fi
+if [ $1 = "c" ]; then
+    rm -f a
+    exit 0;
+fi
+if [ $1 != "b" ]; then
+    echo "Usage: $0 c/b"
+    exit 1;
+fi
+#g++ -o a -std=c++11 -Wall -W -Werror static_assert.cpp
+g++ -o a -std=c++11 -Wall -W -Werror final_override.cpp && ./a
